% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\docType{class}
\name{CellMig}
\alias{CellMig}
\alias{CellMig-class}
\title{The CellMig Class}
\arguments{
\item{object}{An CellMig object, which is a trajectory data frame organized into four columns: cell ID, X coordinates, Y coordinates and Track number, which is the track's path order.}
}
\description{
The CellMig class is the central object storing all information for both random migration (RM) and wound scratch assay (WSA).
It comprises 14 slots
}
\section{Slots}{

\describe{
\item{\code{trajdata}}{The raw trajectory data matrix organized into four columns: cell ID, X coordinates, Y coordinates and Track number, which is the track's path order.}

\item{\code{adjDS}}{A data frame of the trajectory data passed from the WSAprep function.}

\item{\code{cellpos}}{A binary vector showing on which side of the wound cells are located. "0" reffers to a cell located above the wound whereas "1" reffers to a cell located below the wound.}

\item{\code{parE}}{A numeric vector contains estimations for the imageH, woundH, upperE and lowerE.}

\item{\code{preprocessedDS}}{list object of data frames, each data frame shows the trajectories of a single cell.}

\item{\code{DRtable}}{A data frame of the results of running the DiRatio() function.}

\item{\code{MSDtable}}{A data frame of the results of running the MSD() function.}

\item{\code{PerAanSpeedtable}}{A data frame of the results of running the PerAndSpeed() function.}

\item{\code{DACtable}}{A data frame of the results of running the DiAutoCor() function.}

\item{\code{VACtable}}{A data frame of the results of running the VeAutoCor() function.}

\item{\code{ForMigtable}}{A data frame of the results of running the ForwardMigration() function.}

\item{\code{FMItable}}{A data frame of the results of running the FMI() function.}

\item{\code{results}}{A data frame of all the results.}

\item{\code{parCor}}{A data frame for Parameters Correlation.}
}}

\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
