% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{CellMigPCA}
\alias{CellMigPCA}
\title{PCA}
\usage{
CellMigPCA(object, ExpName = "ExpName", parameters = c(1, 2, 3))
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data.}

\item{parameters}{A numeric vector contains the parameters to be included in the Principal Component Analysis. These numbers can be obtained from the outcome of the FinRes() function.}
}
\value{
PCA Graph of cells and PCA Graph of variables.
}
\description{
The CellMigPCA function automatically generates Principal Component Analysis.
}
\examples{
\dontrun{
data(WSADataset)
wasDF=WSADataset[1:1000,]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=95)
wsaTD <-FMI(wsaTD,TimeInterval=10,ExpName="ExpName")
wsaTD <-ForwardMigration(wsaTD,TimeInterval=10,ExpName="ExpName")
wsaTD <-FinRes(wsaTD,ExpName="ExpName",ParCor=FALSE)
PCAplot<-CellMigPCA(wsaTD,parameters=c(1,2))
}

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
