% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{CellMigPCAclust}
\alias{CellMigPCAclust}
\title{PCA Clusters}
\usage{
CellMigPCAclust(object, parameters = c(1, 2, 3), export = FALSE)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data
frames resulted from the PreProcessing.}

\item{parameters}{A numeric vector contains the parameters to be
included in the Principal Component Analysis. These numbers can
be obtained from the outcome of the FinRes() function.}

\item{export}{if `TRUE` (default), exports function output to
CSV file}
}
\value{
PCA Graph of cells and PCA Graph of variables.
}
\description{
The CellMigPCAclust function automatically generates
clusters based on the Principal Component Analysis.
}
\examples{
# Please run this example in an interactive session!
\dontrun{
data(WSADataset)
wasDF=WSADataset[1:300,]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=55)
wsaTD <-FMI(wsaTD,TimeInterval=10)
wsaTD <-ForwardMigration(wsaTD,TimeInterval=10)
wsaTD <-FinRes(wsaTD,ParCor=FALSE)
PCAclust<-CellMigPCAclust(wsaTD,parameters=c(1,9))
}

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
