% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMigRation.R
\docType{package}
\name{CellMigRation}
\alias{CellMigRation}
\title{CellMigRation package}
\description{
Contains 15 trajectory analysis functions.
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
