% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{CellTrackerMainLoop}
\alias{CellTrackerMainLoop}
\title{Cell Tracker Main Loop}
\usage{
CellTrackerMainLoop(
  FinalImage,
  threads,
  tc_obj,
  min_frames_per_cell,
  track_params
)
}
\arguments{
\item{FinalImage}{list of numeric matrices}

\item{threads}{numeric, number of cores to use}

\item{tc_obj}{trackingCell object}

\item{min_frames_per_cell}{numeric, minimum number of frames per cell}

\item{track_params}{a list of tracking parameters}
}
\value{
list of processed data
}
\description{
Tool for Cell Tracker Main Loop
}
\details{
This is an internal function supporting the CellTracker function.
}
\examples{
cellmigRation:::CellTrackerMainLoop(list(1), 1, 1, list(1))


}
\keyword{internal}
