% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{DetectRadii}
\alias{DetectRadii}
\title{Detect Linear Paricle Diameters}
\usage{
DetectRadii(x)
}
\arguments{
\item{x}{numeric or logical vector}
}
\value{
data.frame including two columns: MPOS indicates the centroid position of a particle,
and LEN indicates the diameter size
}
\description{
Estimates the diameters of particles in a numeric or logical vector
}
\examples{
CellMigRation:::DetectRadii(c(0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1))

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
