% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/DiAutoCor.R
\name{DiAutoCor}
\alias{DiAutoCor}
\title{Direction AutoCorrelation}
\usage{
DiAutoCor(
  object,
  TimeInterval = 10,
  ExpName = "ExpName",
  sLAG = 0.25,
  sPLOT = TRUE,
  aPLOT = TRUE
)

DiAutoCor(
  object,
  TimeInterval = 10,
  ExpName = "ExpName",
  sLAG = 0.25,
  sPLOT = TRUE,
  aPLOT = TRUE
)
}
\arguments{
\item{object}{A trajectory data frame organized into four columns: cell ID, X coordinates, Y coordinates and Track number, which is the track's path order.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{sLAG}{A numeric value to be used to get the number of lags for the slope fitting. Default is 0.25, which represents 25 percent of the steps.}

\item{sPLOT}{A logical vector that allows generating individual plots showing the angular persistence across several sequantial time intervals. Default is TRUE.}

\item{aPLOT}{A logical vector that allows generating a plot showing the angular persistence across several sequantial time intervals of all cells. Default is TRUE.}
}
\value{
An CellMig class Object with a data frame and plots. The data frame, which contains six rows: "Cell Number", "Angular Persistence", "Intercept of DA quadratic model","Mean Direction AutoCorrelation (all lags)", "Stable Direction AutoCorrelation through the track" and "Difference between Mean DA and Intercept DA".

Plots and a data frame, which contains six rows: "Cell Number", "Angular Persistence", "Intercept of DA quadratic model","Mean Direction AutoCorrelation (all lags)", "Stable Direction AutoCorrelation through the track" and "Difference between Mean DA and Intercept DA".
}
\description{
The DiAutoCor function automatically compute the angular persistence across several sequantial time intervals.

The DiAutoCor function automatically compute the angular persistence across several sequantial time intervals.
}
\details{
Method DiAutoCor

Method DiAutoCor
}
\examples{
\dontrun{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
rmTD <- rmPreProcessing(rmTD)
rmTD <- DiAutoCor(rmTD,TimeInterval=10,ExpName="ExpName",sLAG=0.25,sPLOT=TRUE,aPLOT=TRUE)
}
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
dac<-DiAutoCor(prepro,TimeInterval=10,ExpName="ExpName",sLAG=0.25,sPLOT=TRUE,aPLOT=TRUE)
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
