% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{DiRatio}
\alias{DiRatio}
\title{Directionality Table}
\usage{
<<<<<<< HEAD
DiRatio(object, TimeInterval = 10, export = FALSE, ExpName = NULL)
=======
DiRatio(object, TimeInterval = 10, ExpName = "ExpName", export = FALSE)
>>>>>>> shiny-page-4
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

<<<<<<< HEAD
\item{export}{if `TRUE` (default), exports function output to CSV file}

\item{ExpName}{string}
}
\value{
An CellMig class object with a data frame stored in the DRtable slot. It contains nine rows: "Cell Number","Directionality Ratio","Mean Cumulative Directionality Ratio","Stable Directionality Ratio", "Number of returns","Min CumDR","Location of Min CumDR, Steps with less CumDR than DR","Directional Persistence"
=======
\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{export}{if `TRUE` (default), exports function output to CSV file}
}
\value{
An CellMig class object with a data frame stored in the DRtable slot. It contains nine rows: "Cell Number","Directionality Ratio","Mean Cumulative Directionality Ratio","Stable Directionality Ratio", "Number of returns","Min CumDR","Location of Min CumDR, Steps with less CumDR than DR","Directional Persistence".
>>>>>>> shiny-page-4
}
\description{
Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straight line length between the start point and the endpoint of the migration trajectory,
}
\details{
Directionality Ratio and Directional persistence
}
\examples{
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
<<<<<<< HEAD
rmTD <- DiRatio(rmTD, export=FALSE)
=======
rmTD <- DiRatio(rmTD, ExpName="Test", export=FALSE)
>>>>>>> shiny-page-4
}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
