% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{FianlizeOptiParams}
\alias{FianlizeOptiParams}
\title{Finalize Output in Parameter Optimization}
\usage{
FianlizeOptiParams(all_results, all_params, target_cell_num, plot)
}
\arguments{
\item{all_results}{list, including all intermediates}

\item{all_params}{data.frame, including all parameter combinations to test}

\item{target_cell_num}{numeric, number of expected cells}

\item{plot}{logical, shall a series of plots be generated}
}
\value{
a list including test results; an empty list
is returned if an error is encountered.
}
\description{
Finalize Output as part of the Optimization Parameter process
}
\details{
This is an internal function supporting the Optimization
Parameter process
}
\examples{
cellmigRation:::FianlizeOptiParams(list(1), data.frame(1), 10, FALSE)

}
\keyword{internal}
