% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixExpName}
\alias{fixExpName}
\title{Handle non-NULL ExpName}
\usage{
fixExpName(x)
}
\arguments{
\item{x}{string, name of the experiment.}
}
\value{
A string referring to the adjusted experiment name.
}
\description{
The fixExpName helps adjusting the name of the experiment
in case it is not NULL.
}
\examples{
cellmigRation:::fixExpName("Hello World")

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
