% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixFM2}
\alias{fixFM2}
\title{Forward Migration Second Part}
\usage{
fixFM2(Object, Step)
}
\arguments{
\item{Step}{A numeric value of the number of trajectory steps.}

\item{object}{\code{CellMig} class object, which is a list of data
frames resulted from the PreProcessing.}
}
\value{
An CellMig class Object.
}
\description{
This function is a part of the ForwardMigration function,
which generates data and plots for forward persistence and speed.
}
\examples{
cellmigRation:::fixFM2(1, 1)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
