% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixFM5}
\alias{fixFM5}
\title{Forward Migration Fifth Part}
\usage{
fixFM5(Object, TimeInterval, Step)
}
\arguments{
\item{TimeInterval}{A numeric value of the time elapsed between
successive frames in the time-lapse stack.}

\item{Step}{A numeric value of the number of trajectory steps.}

\item{object}{\code{CellMig} class object, which is a list of data
frames resulted from the PreProcessing.}
}
\value{
A data frame named "FMResultsTable".
}
\description{
This function is a part of the ForwardMigration function,
which generates data and plots for forward persistence and speed.
}
\examples{
cellmigRation:::fixFM5(1, 1, 1)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
