% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixFM6}
\alias{fixFM6}
\title{Forward Migration Sixst Part}
\usage{
fixFM6(
    Object,
    FMResultsTable,
    Step,
    sfptPLOT,
    afptPLOT,
    export,
    color,
    TimeInterval,
    sfpPLOT,
    ExpName,
    new.fld
)
}
\arguments{
\item{FMResultsTable}{A data frame resulted from the fixFM6().}

\item{Step}{A numeric value of the number of trajectory steps.}

\item{sfptPLOT}{A logical vector that allows generating individual
plots of persistence time vs speed per cell. Default is TRUE.}

\item{afptPLOT}{A logical vector that allows generating a plot of
persistence time vs speed for all cells. Default is TRUE.}

\item{export}{if `TRUE` (default), exports function output to CSV
file}

\item{color}{A vector of colors that will be used for the plots}

\item{TimeInterval}{A numeric value of the time elapsed between
successive frames in the time-lapse stack.}

\item{sfpPLOT}{A logical vector that allows generating individual
plots of angular persistence vs speed per cell. Default is TRUE.}

\item{ExpName}{String, name of the experiment}

\item{new.fld}{path to the folder where to save files}

\item{object}{\code{CellMig} class object, which is a list of data
frames resulted from the PreProcessing.}
}
\value{
A data frame named "FMResultsTable".
}
\description{
This function is a part of the ForwardMigration function,
which generates data and plots for forward persistence and speed.
}
\examples{
cellmigRation:::fixFM6(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
