% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixID4}
\alias{fixID4}
\title{Pre-processing Fourth Part}
\usage{
fixID4(ID_split)
}
\arguments{
\item{ID_split}{A list of data frames.}

\item{TimeInterval}{A numeric value of the time elapsed between}
}
\value{
A list of data frames.
}
\description{
This function prepare the data in each data frame
as a part of the pre-processing.
}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
