% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixPER2}
\alias{fixPER2}
\title{Persistence and Speed Second Part}
\usage{
fixPER2(
  Object,
  PerResultsTable,
  PtSplot,
  AllPtSplot,
  export,
  color,
  TimeInterval,
  ExpName,
  new.fld
)
}
\arguments{
\item{TimeInterval}{A numeric value of the time elapsed between}

\item{ExpName}{String, name of the experiment}

\item{new.fld}{path to the folder where to save files}

\item{x}{\code{CellMig} class object, which is a list of data}
}
\value{
A data frame named "PerResultsTable".
}
\description{
This function is a part of the PerAndSpeed(), which
generates data and plots for persistence and speed.
}
\examples{
cellmigRation:::fixPER2(1, 1, 1, 1, 1, 1, 1, 1, 1)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
