% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_internal_v3.R
\name{fixPER3}
\alias{fixPER3}
\title{Persistence and Speed Third Part}
\usage{
fixPER3(
  Object,
  PerResultsTable,
  ApSplot,
  AllApSplot,
  export,
  color,
  TimeInterval,
  ExpName,
  new.fld
)
}
\arguments{
\item{Object}{\code{CellMig} class object, which is a list of data.}

\item{PerResultsTable}{A data frame.}

\item{ApSplot}{A logical vector that allows generating individual
plots of angular persistence vs speed per cell. Default is TRUE.}

\item{AllApSplot}{A logical vector that allows generating a plot
of angular persistence vs speed of all cells. Default is TRUE.}

\item{export}{if `TRUE` (default), exports function output.}

\item{color}{A vector of colors that will be used for the plots}

\item{TimeInterval}{A numeric value of the time elapsed between}

\item{ExpName}{String, name of the experiment}

\item{new.fld}{path to the folder where to save files}
}
\value{
A data frame named "PerResultsTable".
}
\description{
This function is a part of the PerAndSpeed(), which
generates data and plots for persistence and speed.
}
\examples{
cellmigRation:::fixPER3(1,1,1,1,1,1,1,1,1)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
\keyword{internal}
