% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{GenAllCombos}
\alias{GenAllCombos}
\title{Generate All Combinations
 
Generate All Combinations as part of the Optimization Parameter process}
\usage{
GenAllCombos(...)
}
\arguments{
\item{...}{a series of arguments where each argument is a vector of values 
to be combined together}
}
\value{
a data frame of combined parameters to be tested
}
\description{
Generate All Combinations
 
Generate All Combinations as part of the Optimization Parameter process
}
\details{
This is an internal function supporting the Optimization Parameter 
steps
}
\examples{
cellmigRation:::GenAllCombos(A=c(1,2,3), B = 10, C = c("x", "y", "z"))

}
\keyword{internal}
