% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getCellImages}
\alias{getCellImages}
\alias{getCellImages,trackedCells-method}
\title{Method getCellImages}
\usage{
getCellImages(x)

\S4method{getCellImages}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a list including all images
}
\description{
Retrieve Images from a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
getCellImages(TrackCellsDataset)

}
