% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{getCellsMeta}
\alias{getCellsMeta}
\title{Get MetaData}
\usage{
getCellsMeta(tc_obj)
}
\arguments{
\item{tc_obj}{a trackedCells object}
}
\value{
a list including four items: tiff filename, experiment name, condition label, and replicate ID.
}
\description{
Extract MetaData from a trackedCells object
}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
