% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{getCellTrackMeta}
\alias{getCellTrackMeta}
\alias{getCellTrackMeta,trackedCells-method}
\title{Method getCellTrackMeta}
\usage{
getCellTrackMeta(x)

\S4method{getCellTrackMeta}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\description{
Retrieve Metadata from a \code{trackedCells} object.
}
