% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getCellTracks}
\alias{getCellTracks}
\alias{getCellTracks,trackedCells-method}
\title{Method getCellTracks}
\usage{
getCellTracks(x)

\S4method{getCellTracks}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a data.frame including Cell Tracks
}
\description{
Retrieve Cell Tracks from a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
getCellTracks(TrackCellsDataset)

}
