% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{getFMItable}
\alias{getFMItable}
\title{Getting the Forward Migration Index}
\usage{
getFMItable(object)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of
data frames resulted from the PreProcessing.}
}
\value{
A data frame for the FMI.
}
\description{
The FMI function automatically generates data
for the forward migration index
}
\examples{
data(WSADataset)
wasDF=WSADataset[1:300,]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=55)
wsaTD <-FMI(wsaTD,TimeInterval=10, export=FALSE)
head(getFMItable(wsaTD))

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
