% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{getImageCentroids}
\alias{getImageCentroids}
\alias{getImageCentroids,trackedCells-method}
\title{Method getImageCentroids}
\usage{
getImageCentroids(x)

\S4method{getImageCentroids}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\description{
Retrieve Image Centroids from a \code{trackedCells} object.
}
