% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_02.R
\name{getImageStacks}
\alias{getImageStacks}
\title{Get Image Stacks}
\usage{
getImageStacks(tc_obj)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}
}
\value{
a list including stack images (formatted as numeric matrices)
}
\description{
Extract Images Stacks from a trackedCells object
}
\examples{
x0 <- cellmigRation::TrackCellsDataset
y0 <- getImageStacks(x0)
graphics::image(y0[[1]][200:400, 100:300])

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
