% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{getOptimizedParameters}
\alias{getOptimizedParameters}
\title{Get Auto Optimized Parameters}
\usage{
getOptimizedParameters(tc_obj)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}
}
\value{
a list including optimized parameter values
(lnoise, diameter, and threshold)
}
\description{
Extract Parameters that were automatically optimized
}
\examples{
x <- get(data(TrackCellsDataset))
getOptimizedParameters(x)


}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
