% Manually Edited
% Expected to be identical to the corresponding
% Rd file Generated by roxygen2
\name{getOptimizedParams}
\alias{getOptimizedParams}
\title{Get Auto Optimized Parameters}
\usage{
getOptimizedParams(tc_obj)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}
}
\value{
a list including optimized parameter values (lnoise, diameter, and threshold)
}
\description{
Extract Parameters that were automatically optimized
}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
