% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getProcessedImages}
\alias{getProcessedImages}
\alias{getProcessedImages,trackedCells-method}
\title{Method getProcessedImages}
\usage{
getProcessedImages(x)

\S4method{getProcessedImages}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a list including all processed images
}
\description{
Retrieve Processed Images from a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
getProcessedImages(TrackCellsDataset)

}
