% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{getResults}
\alias{getResults}
\title{Final Results}
\usage{
getResults(object)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data
frames resulted from the PreProcessing.}
}
\value{
A data frame that contains all the results.
}
\description{
The FinRes function automatically generates a data
frame that contains all the results.
}
\examples{
data(WSADataset)
wasDF <- WSADataset[1:300, ]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=55)
wsaTD <-FMI(wsaTD,TimeInterval=10)
wsaTD <-ForwardMigration(wsaTD,TimeInterval=10,)
wsaTD <-FinRes(wsaTD,ParCor=FALSE, export=FALSE)
head(getResults(wsaTD))

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
