% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{getTracks}
\alias{getTracks}
\title{Get Track Data}
\usage{
getTracks(tc_obj, attach_meta = FALSE)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}

\item{attach_meta}{logical, shall metaData be attached to tracks}
}
\value{
a data.frame including cell tracks data
}
\description{
Extract Track Data from a trackedCells object
}
\examples{
<<<<<<< HEAD
message("this example may take up to several mins to complete")
\dontrun{
x <- cellmigRation::TrackCellsDataset
x <- OptimizeParams(x)
x <- CellTracker(x)
getTracks(x)[1:10,]
}
=======
x <- cellmigRation::TrackCellsDataset
getTracks(x)[seq(1,10,by=1),]
>>>>>>> BiocNewVersion


}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
