% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{initializeTrackParams}
\alias{initializeTrackParams}
\title{Initialize Tracking parameters}
\usage{
initializeTrackParams(dd = 0, params)
}
\arguments{
\item{dd}{numeric, value of the dd param}

\item{params}{a list containing a few tracking parameters that are
needed for the analysis}
}
\value{
a list including parsed arguments
}
\description{
Initialize parameter variables used for the tracking
}
\examples{
cellmigRation:::initializeTrackParams(0, NULL)


}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
