% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{innerBondRaster}
\alias{innerBondRaster}
\title{Inner Bond Raster}
\usage{
innerBondRaster(xyzs, maxdisp, j = 1, env)
}
\arguments{
\item{xyzs}{data.frame, including input cell centroid positions}

\item{maxdisp}{numeric, value of maximum cell dispersion in pixels}

\item{env}{environment, including all objects used for the tracking}

\item{i}{numeric, index of the iteration cycle}
}
\value{
FALSE is returned; objects in env are updated
}
\description{
Perform Inner Bond Raster as part of the Cell Tracking Processing
}
\details{
a message is printed if an issue (typically arising by a 
non-suitable environment being passed as the env argument)  is detected. 
See the example below.
}
\examples{
cellmigRation:::innerBondRaster(data.frame(1), 1, 1, new.env())


}
\keyword{internal}
