% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/MSD.R
\name{MSD}
\alias{MSD}
\title{Mean Square Displacement}
\usage{
MSD(
  object,
  TimeInterval = 10,
  ExpName = "ExpName",
  sLAG = 0.25,
  ffLAG = 0.25,
  SlopePlot = TRUE,
  AllSlopesPlot = TRUE,
  FurthPlot = TRUE,
  AllFurthPlot = TRUE
)

MSD(
  object,
  TimeInterval = 10,
  ExpName = "ExpName",
  sLAG = 0.25,
  ffLAG = 0.25,
  SlopePlot = TRUE,
  AllSlopesPlot = TRUE,
  FurthPlot = TRUE,
  AllFurthPlot = TRUE
)
}
\arguments{
\item{object}{A trajectory data frame organized into four columns: cell ID, X coordinates, Y coordinates and Track number, which is the track's path order.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{sLAG}{A numeric value to be used to get the number of lags for the slope fitting. Default is 0.25, which represents 25 percent of the steps.}

\item{ffLAG}{A numeric value to be used to get the number of lags for the  Furth formula fitting. Default is 0.25, which represents 25 percent of the steps.}

\item{SlopePlot}{A logical vector that allows generating individual plots showing the slope of the mean square displacement of the movement of individual cells. Default is TRUE.}

\item{AllSlopesPlot}{A logical vector that allows generating a plot showing the slope of the mean square displacement of the movement of all cells. Default is TRUE.}

\item{FurthPlot}{A logical vector that allows generating individual plots fitting the Furth formula using generalized regression by the Nelder–Mead method simplex method per cell. Default is TRUE.}

\item{AllFurthPlot}{A logical vector that allows generating a plot fitting the Furth formula using generalized regression by the Nelder–Mead method simplex method for all cells. Default is TRUE.}
}
\value{
An CellMig class object with a data frame and plots. The data frame is stored in the MSDtable slot.

A data frame and plots
}
\description{
The MSD function automatically compute the mean square displacements across several sequantial time intervals. MSD parameters are used to assess the area explored by cells over time.

The MSD function automatically compute the mean square displacements across several sequantial time intervals. MSD parameters are used to assess the area explored by cells over time.
}
\details{
Method msd

Method msd
}
\examples{
\dontrun{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
rmTD <- rmPreProcessing(rmTD)
rmTD<-MSD(rmTD,sLAG=0.25, ffLAG=0.25)
}
\dontrun{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
msd<-MSD(prepro,sLAG=0.25, ffLAG=0.25)
}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
