% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{NextOdd}
\alias{NextOdd}
\title{Return the Next Odd Integer}
\usage{
NextOdd(x)
}
\arguments{
\item{x}{a vector of class numeric}
}
\value{
a vector of class integer
}
\description{
Returns the smallest odd number bigger than the number(s) provided as the argument
}
\examples{
cellmigRation:::NextOdd(2:5)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
