% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{nontrivialBondTracking}
\alias{nontrivialBondTracking}
\title{Non-Trivial Bond Tracking}
\usage{
nontrivialBondTracking(xyzs, maxdisp, env = eee)
}
\arguments{
\item{xyzs}{data.frame, including input cell centroid positions}

\item{maxdisp}{numeric, value of maximum cell dispersion in pixels}

\item{env}{environment, including all objects used for the tracking}
}
\value{
FALSE is returned; objects in env are updated
}
\description{
Perform Non-Trivial Bond Tracking as part of the Cell Tracking Processing
}
\details{
a message is printed if an issue (typically arising by a 
non-suitable environment being passed as the env argument)  is detected. 
See the example below.
}
\examples{
cellmigRation:::nontrivialBondTracking(data.frame(1), 1, new.env())

}
\keyword{internal}
