% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{OptimizeParamsMainLoop}
\alias{OptimizeParamsMainLoop}
\title{Main Loop to Parameter Optimization}
\usage{
OptimizeParamsMainLoop(
  tmp_img = tmp_img,
  all_params = all_params,
  threads = threads,
  verbose = verbose
)
}
\arguments{
\item{tmp_img}{numeric matrix, corresponding to the}

\item{all_params}{data.frame, including all parameter combinations to test}

\item{threads}{numeric, number of cores to use for the parallelization}

\item{verbose}{logical, shall info be printed to console}
}
\value{
a list including test results; an empty list
is returned if an error is encountered.
}
\description{
Main Loop as part of the Optimization Parameter process
}
\details{
This is an internal function supporting the Optimization
Parameter process
}
\examples{
cellmigRation:::OptimizeParamsMainLoop(matrix(1), data.frame(1), 1, FALSE)

}
\keyword{internal}
