% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{Parallel4OptimizeParams}
\alias{Parallel4OptimizeParams}
\title{Paralle Parameter Optimization}
\usage{
Parallel4OptimizeParams(
  tmp_img = tmp_img,
  all_params = all_params,
  use.cores = use.cores,
  verbose = verbose
)
}
\arguments{
\item{tmp_img}{numeric matrix, corresponding to the}

\item{all_params}{data.frame, including all parameter combinations to test}

\item{use.cores}{numeric, number of cores to use for the parallelization}

\item{verbose}{logical, shall info be printed to console}
}
\value{
a list including test results; an empty list
is returned if an error is encountered.
}
\description{
Parallel Optimization as part of the Optimization Parameter process
}
\details{
This is an internal function supporting the Optimization
Parameter process
}
\examples{
cellmigRation:::Parallel4OptimizeParams(matrix(1), data.frame(1), 1, TRUE)

}
\keyword{internal}
