% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{ParallelTrackLoop}
\alias{ParallelTrackLoop}
\title{Multi Core Tracking Loop}
\usage{
ParallelTrackLoop(FinalImage, use.cores, min_frames_per_cell, track_params)
}
\arguments{
\item{FinalImage}{a list of numeric matrices (images)}

\item{use.cores}{numeric, number of cores to use}

\item{min_frames_per_cell}{numeric, minimum number of frames}

\item{track_params}{a list of tracking parameters}
}
\value{
list of processed data
}
\description{
Tool for Multi Core Tracking Loop
}
\details{
This is an internal function supporting the CellTracker function.
}
\examples{
cellmigRation:::ParallelTrackLoop(list(), 1, 1, list())



}
\keyword{internal}
