% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_02.R
\name{pkfnd}
\alias{pkfnd}
\title{Find Signal Peaks}
\usage{
pkfnd(im, th, sz = NULL)
}
\arguments{
\item{im}{image to process, particle should be bright spots on dark background
with little noise ofen an bandpass filtered brightfield image}

\item{th}{the minimum brightness of a pixel that might be local maxima.
NOTE: Make it big and the code runs faster but you might miss some particles.
Make it small and you'll get everything and it'll be slow.}

\item{sz}{if your data is noisy, (e.g. a single particle has multiple local maxima),
then set this optional keyword to a value slightly larger than the diameter of your blob.
If multiple peaks are found withing a radius of sz/2 then the code will keep only the brightest.
Also gets rid of all peaks within sz of boundary}
}
\value{
a numeric data.frame with two columns, with the coordinates of local maxima
}
\description{
Finds local maxima in an image to pixel level accuracy. This provides a rough guess
of particle centers to be used by cntrd(). Inspired by the lmx subroutine of
Grier and Crocker's. CREATED: Eric R. Dufresne, Yale University, Feb 4 2005.
}
\examples{
message("the following example may take up to one minute to run")
\dontrun{
x0 <- cellmigRation::TrackCellsDataset@images$images[[1]][100:500, 200:600]
b <- cellmigRation:::bpass(image_array = x0,
                           lnoise = 1, 
                           lobject = 15,
                           threshold = 10) 
pk <- cellmigRation:::pkfnd(b, 10, 17)
cnt <- cellmigRation:::cntrd(im = b, mx = pk, sz = 17)
cnt[1:10,]
}

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
