% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{plot3DTracks}
\alias{plot3DTracks}
\title{A 3D rose-plot}
\usage{
plot3DTracks(object, VS = 3, size = 2, cells)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{VS}{A numeric value of the vertical separator between cells.}

\item{size}{A numeric value of the point's size.}

\item{cells}{A numeric vector containing the cell's numbers to be plotted.}
}
\value{
A 3D rose-plot showing the tracks of particular cells.
}
\description{
Plotting the trajectory data of particular cells in 3D.
}
\details{
The 3D visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(Trajectory_dataset)
rmDF=Trajectory_dataset[1:2000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
plot3DTracks(rmTD, VS=3, size=2,cells=c(1,10,15))

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
