% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{plotAllTracks}
\alias{plotAllTracks}
\title{A 2D rose-plot}
\usage{
plotAllTracks(
<<<<<<< HEAD
  object,
  Type = "l",
  FixedField = TRUE,
  export = FALSE,
  ExpName = NULL
=======
    object,
    Type = "l",
    FixedField = TRUE,
    export = FALSE,
    ExpName = NULL
>>>>>>> BiocNewVersion
)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list
of data frames resulted from the PreProcessing.}

\item{Type}{has to be one of the following: c("p", "l", "b", "o")
<<<<<<< HEAD
"p": Points
"l": Lines
"b": Both
"o": Both "overplotted"}
=======
"p": Points;
"l": Lines;
"b": Both;
"o": Both "overplotted".}
>>>>>>> BiocNewVersion

\item{FixedField}{logical(1) Allows generating a plot with
fixed field 800um x 800um. Default is TRUE.}

\item{export}{if `TRUE` (default), exports plot to JPG file}

\item{ExpName}{string, name of the experiment. Can be NULL}
}
\value{
A 2D rose-plot showing the tracks of all cells.
}
\description{
Plotting the trajectory data of all cells.
}
\details{
The visualization shows centered trajectories where
the starting point of each track is located at the origin
of the coordinate system (X=0,Y=0).
}
\examples{
<<<<<<< HEAD
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
plotAllTracks(rmTD,Type="l", FixedField=TRUE,export=FALSE)
=======
rmTD <- cellmigRation::preProcCellMig
plotAllTracks(object=rmTD, Type="l", FixedField=TRUE,export=FALSE)
>>>>>>> BiocNewVersion

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
