% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{PostProcessTracking}
\alias{PostProcessTracking}
\title{Post Processing Tracking Routine 
 
Tool for Post Processing Tracking Routine}
\usage{
PostProcessTracking(tc_obj, all_centroids, track_params)
}
\arguments{
\item{tc_obj}{a trackedCells object}

\item{all_centroids}{list, frame by frame centroids}

\item{track_params}{a list of tracking parameters}
}
\value{
list of processed data
}
\description{
Post Processing Tracking Routine 
 
Tool for Post Processing Tracking Routine
}
\details{
This is an internal function supporting the CellTracker function.
}
\examples{
cellmigRation:::PostProcessTracking(list(1), list(2), list(A=3))



}
\keyword{internal}
