% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{preProcCellMig}
\alias{preProcCellMig}
\title{Trajectories of 11 cells}
\format{
a list including 21 elements
}
\usage{
data(preProcCellMig)
}
\description{
Intermediates and Results from Cell Tracking Analyses,
used as a representative example of a S4 CellMig object
}
\examples{
data(preProcCellMig)

}
\keyword{internal}
