% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{PreProcessing}
\alias{PreProcessing}
\title{Method preprocessing}
\usage{
PreProcessing(object, PixelSize, TimeInterval)
}
\arguments{
\item{object}{A trajectory data frame organized into four columns: cell ID, X coordinates, Y coordinates and Track number, which is the track's path order.}

\item{PixelSize}{A numeric value of the physical size of a pixel.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}
}
\value{
A list of data frames, each data frame shows the trajectories of a single cell.
}
\description{
The PreProcessing function automatically adjusts the X and Y coordinates based on the physical size of the pixel “PixelSize” and let all cells start from a single point (X=0,Y=0). Furthermore, it computes several descriptive parameters, including distances, absolute angles,relative angles, persistence time, squared speed, cumulative distance, Directional ratio and acceleration.
}
\examples{
data(Trajectory_dataset)
df<-Trajectory_dataset
preproc<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
