% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{rmPreProcessing}
\alias{rmPreProcessing}
\title{Data preprocessing for random migration (RM)}
\usage{
rmPreProcessing(object, PixelSize = 1.24, TimeInterval = 10)
}
\arguments{
\item{object}{\code{CellMig} class object.}

\item{PixelSize}{A numeric value of the physical size of a pixel.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}
}
\value{
An CellMig class object with preprocessed data.
}
\description{
This function allows preprocessing of the trajectory data from random migration (RM) experiments.
}
\examples{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
\dontrun{
rmTD <- rmPreProcessing(rmTD)
}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
