% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setCellMigSlot}
\alias{setCellMigSlot}
\alias{setCellMigSlot,CellMig,character-method}
\alias{setCellMigSlot,CellMig,character,ANY-method}
\title{Method setCellMigSlot}
\usage{
setCellMigSlot(x, slot, value)

\S4method{setCellMigSlot}{CellMig,character}(x, slot, value)
}
\arguments{
\item{x}{a \code{CellMig}-class object}

\item{slot}{string pointing to the slot to be updated}

\item{value}{ANY value to be written}
}
\value{
a CellMig object
}
\description{
Set Data of a slot in a \code{CellMig} object.
}
\examples{
data("TrajectoryDataset")
x <- CellMig(TrajectoryDataset)
setCellMigSlot(x, "cellpos", c(1, 2, 3))


}
