% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setCellTracks}
\alias{setCellTracks}
\alias{setCellTracks,trackedCells,matrix-method}
\title{Method setCellTracks}
\usage{
setCellTracks(x, tracks)

\S4method{setCellTracks}{trackedCells,matrix}(x, tracks)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{tracks}{a matrix including all cell tracks}
}
\value{
a trackedCells object
}
\description{
Set Tracks of a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
setCellTracks(TrackCellsDataset, matrix())

}
