% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{setExpName}
\alias{setExpName}
\alias{setExpName,CellMig,character-method}
\title{Method setExpName}
\usage{
setExpName(x, ExpName)

\S4method{setExpName}{CellMig,character}(x, ExpName)
}
\arguments{
\item{x}{a \code{CellMig}-class object}

\item{ExpName}{string corresponding to the ExpName}
}
\description{
Set Experiment Name of a \code{CellMig} object.
}
