% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setProcessedImages}
\alias{setProcessedImages}
\alias{setProcessedImages,trackedCells,list-method}
\title{Method setProcessedImages}
\usage{
setProcessedImages(x, procImages)

\S4method{setProcessedImages}{trackedCells,list}(x, procImages)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{procImages}{a list including all metadata}
}
\value{
a trackedCells object
}
\description{
Set Processed Images of a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
prc.im <- getProcessedImages(TrackCellsDataset)
setProcessedImages(TrackCellsDataset, prc.im)


}
