% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{setTrackingStats}
\alias{setTrackingStats}
\alias{setTrackingStats,trackedCells-method}
\alias{setTrackingStats,trackedCells,ANY,ANY-method}
\title{Method setTrackingStats}
\usage{
setTrackingStats(x, population, cells)

\S4method{setTrackingStats}{trackedCells}(x, population, cells)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{population}{population-level statistics}

\item{cells}{cell-level statistics}
}
\description{
Set Tracking Statistics of a \code{trackedCells} object.
}
