% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{subNetworkTracking}
\alias{subNetworkTracking}
\title{Subnetwork Tracking}
\usage{
subNetworkTracking(xyzs, maxdisp, env)
}
\arguments{
\item{xyzs}{data.frame, including input cell centroid positions}

\item{maxdisp}{numeric, value of maximum cell dispersion in pixels}

\item{env}{environment, including all objects used for the tracking}
}
\value{
FALSE is returned while objects in env are updated
}
\description{
Perform Internal Subnetwork Tracking as part of the Cell Tracking Processing
}
\details{
a message is printed if an issue (typically arising by a
non-suitable environment being passed as the env argument)  is detected.
See the example below.
}
\examples{
cellmigRation:::subNetworkTracking(data.frame(1), 1, new.env())

}
\keyword{internal}
