% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra_v2.R
\docType{data}
\name{ThreeConditions}
\alias{ThreeConditions}
\title{Intermediates and Results from Cell Tracking Analyses}
\format{
a list including 3 elements
}
\usage{
data(ThreeConditions)
}
\description{
Intermediates and Results from Cell Tracking Analyses,
used to build the package vignette.
}
\details{
BT549 cell trajectories were computed using cellmigRation.
Imaging experiments were performed as described by Ghannoum S
et al (paper in preparation). Briefly, triple negative breast
cancer BT549 cells were cultured in RPMI supplemented with 10% FCS
and 1% penicillin/streptomycin. Cells were transduced with
NucLight green lentivirus (Essen BioScience), and then sorted by
fluorescence-activated cell sorting (FACS). GFP-positive cells
were seeded at a 1:3 ratio with untransduced BT549 cells in
96-well image-lock plates (EssenBio) at a density of 1000
total cells per well.
Cells were scanned at ten-minute intervals over 24h using an
Incucyte S3 Live-Cell microscope (EssenBio) at 10x magnification
and a Basler Ace 1920-155um camera with CMOS sensor.
Cells were treated with 100 uM Rac1 Inhibitor (1177865-17-6,
Calbiochem) or left untreated (controls).  TIFF images
were imported and processed using the cellmigRation library.
}
\examples{
data(ThreeConditions)

}
\keyword{internal}
