% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{TrackCellsDataset}
\alias{TrackCellsDataset}
\title{Sample Stack of Fluorescent Cells}
\format{
a \code{trackedCells} object including 10 stacks
}
\usage{
data(TrackCellsDataset)
}
\description{
Sample Stack of Fluorescent Cells to be used for
computing cell tracks and stats
}
\details{
This stack of Fluorescent Cell images was obtained by 
processing TIFF files via cellmigRation. 
Imaging experiments were performed as described by Ghannoum S 
et al (paper in preparation). Briefly, triple negative breast 
cancer BT549 cells were cultured in RPMI supplemented with 10% FCS 
and 1% penicillin/streptomycin. Cells were transduced with 
NucLight green lentivirus (Essen BioScience), and then sorted by  
fluorescence-activated cell sorting (FACS). GFP-positive cells 
were seeded at a 1:3 ratio with untransduced BT549 cells in 
96-well image-lock plates (EssenBio) at a density of 1000 
total cells per well.
Cells were scanned at ten-minute intervals over 24h using an 
Incucyte S3 Live-Cell microscope (EssenBio) at 10x magnification 
and a Basler Ace 1920-155um camera with CMOS sensor. 
A small selection of cells and image stacks is included in
this dataset.
}
\examples{
data(TrackCellsDataset)

}
\keyword{internal}
