% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{trackSlideProcessing}
\alias{trackSlideProcessing}
\title{Tracking Slide Processing}
\usage{
trackSlideProcessing(xyzs, maxdisp, i, env)
}
\arguments{
\item{xyzs}{data.frame, including input cell centroid positions}

\item{maxdisp}{numeric, value of maximum cell dispersion in pixels}

\item{env}{environment, including all objects used for the tracking}
}
\value{
FALSE is returned as long as cells are detected and tracked; 
TRUE is returned when no further cells to track are found; 
objects in env are updated
}
\description{
Frame-by-frame Slide Processing as part of the Cell Tracking Processing
}
\details{
a message is printed if an issue (typically arising by a 
non-suitable environment being passed as the env argument)  is detected. 
See the example below.
}
\examples{
cellmigRation:::trackSlideProcessing(data.frame(1), 1, 1, new.env())

}
\keyword{No}
\keyword{Warning-}
\keyword{for}
\keyword{found}
\keyword{internal}
\keyword{positions}
\keyword{t='}
