% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{trackSlideWrapUp}
\alias{trackSlideWrapUp}
\title{Tracking Slide Wrap Up}
\usage{
trackSlideWrapUp(xyzs, maxdisp, i, env)
}
\arguments{
\item{xyzs}{data.frame, including input cell centroid positions}

\item{maxdisp}{numeric, value of maximum cell dispersion in pixels}

\item{i}{integer, index of the current cycle}

\item{env}{environment, including all objects used for the tracking}
}
\value{
FALSE is returned; objects in env are updated
}
\description{
Perform Tracking Slide Wrap Up as part of the Cell Tracking Processing
}
\details{
a message is printed if an issue (typically arising by a
non-suitable environment being passed as the env argument)  is detected.
See the example below.
}
\examples{
cellmigRation:::trackSlideWrapUp(data.frame(1), 1, 1, new.env())

}
\keyword{internal}
