% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trajectory_dataset.R
\docType{data}
\name{Trajectory_dataset}
\alias{Trajectory_dataset}
\title{Trajectories of 350 cells}
\format{A data frame with 50216 rows and 4 columns}
\usage{
data(Trajectory_dataset)
}
\description{
A dataset containing the coordinates and the ID of 350 cells from a dense random migration experiment
}
\examples{
data(Trajectory_dataset)

}
\keyword{datasets}
