% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_datasets.R, R/all_extra.R
\docType{data}
\name{TrajectoryDataset}
\alias{TrajectoryDataset}
\title{Trajectories of 350 cells}
\format{
A data frame with 50216 rows and 4 columns

A data frame with 50216 rows and 4 columns
}
\usage{
data(TrajectoryDataset)

data(TrajectoryDataset)
}
\description{
A dataset containing the coordinates and the ID of 350 cells from a dense random migration experiment

A dataset containing the coordinates and the ID of 350 cells from a dense random migration experiment
}
\examples{
data(TrajectoryDataset)

data(TrajectoryDataset)

}
\keyword{datasets}
