% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{ValidateTrackingArgs}
\alias{ValidateTrackingArgs}
\title{Validate Tracking Arguments}
\usage{
ValidateTrackingArgs(
    import_optiParam_from = NULL,
    lnoise = NULL,
    diameter = NULL,
    threshold = NULL,
    maxDisp = NULL,
    memory_b = 0,
    goodenough = 0,
    show_plots = FALSE,
    verbose = FALSE
)
}
\arguments{
\item{import_optiParam_from}{a trackedCells object, defaults to NULL}

\item{lnoise}{numeric, lnoise value}

\item{diameter}{numeric, lnoise value}

\item{threshold}{numeric, lnoise value}

\item{maxDisp}{numeric, lnoise value}

\item{memory_b}{numeric, should be 0}

\item{goodenough}{numeric, should be 0}

\item{show_plots}{logical, shall plots be shown}

\item{verbose}{logical, shall info be printed to console}
}
\value{
list of processed data
}
\description{
Tool for Validate Tracking Arguments
}
\details{
This is an internal function supporting the CellTracker function.
}
\examples{
cellmigRation:::ValidateTrackingArgs()



}
\keyword{internal}
