% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{VignBuilderDataset}
\alias{VignBuilderDataset}
\title{Intermediates and Results from Cell Tracking Analyses}
\format{
a list including 21 elements
}
\usage{
data(VignBuilderDataset)
}
\description{
Intermediates and Results from Cell Tracking Analyses,
used to build the package vignette.
}
\examples{
data(VignBuilderDataset)

}
\keyword{internal}
