% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{visualizeTrcks}
\alias{visualizeTrcks}
\title{Visualize Cell Tracks}
\usage{
visualizeTrcks(
  tracks,
  width_px,
  height_px,
  i.slice = 1,
  pnt.cex = 1.2,
  lwd = 1.2,
  col = "red"
)
}
\arguments{
\item{tracks}{cell tracks}

\item{width_px}{width in pixels}

\item{height_px}{height in pixels}

\item{i.slice}{index of the stack slice to use}

\item{pnt.cex}{cex for the points (circles) drawn around the cells}

\item{lwd}{lwd of cell tracks}

\item{col}{color used for the cell tracks, .g. "red"}
}
\value{
None
}
\description{
Annotates an image with cell centroids by adding cell ROIs and drawing cell tracks
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
