% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{warnMessage}
\alias{warnMessage}
\title{Print Warning Messages}
\usage{
warnMessage(warn_log, quiet = FALSE)
}
\arguments{
\item{warn_log}{list, including the warning logs}

\item{quiet}{logical, shall the warning be printed}
}
\value{
log of warning messages is returned
}
\description{
Print warning messages to the console when an issue is encountered
}
\examples{
cellmigRation:::warnMessage(list("Hello world"), FALSE)


}
\keyword{internal}
