% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{CellMigPCA}
\alias{CellMigPCA}
\title{PCA}
\usage{
CellMigPCA(object, ExpName = "ExpName", parameters = c(1, 2, 3))
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data.}

\item{parameters}{A numeric vector contains the parameters to be included in the Principal Component Analysis. These numbers can be obtained from the outcome of the FinRes() function.}
}
\value{
PCA Graph of cells and PCA Graph of variables.
}
\description{
The CellMigPCA function automatically generates Principal Component Analysis.
}
\details{
Method PCA
}
\examples{
data(Trajectory_dataset)
wsaTD <- CellMig(WSAdataset)
wsaTD <- wsaPreProcessing(wsaTD)
wsaTD <-FinRes(wsaTD,ExpName="ExpName",ParCor=FALSE)
CellMigPCA(wsaTD,ExpName="ExpName",parameters=c(1,2,3))
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
