% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{package}
\name{cellmigRation-package}
\alias{cellmigRation}
\alias{cellmigRation-package}
\title{Track And Analyze Cell Migrations}
\description{
Track Fluorescent Cells and Analyze their movements.
Compute Migration Statistics and advanced metrics to understand
motility and movement characteristics of a population of cells.
}
\seealso{
Useful links:
\itemize{
<<<<<<< HEAD
  \item \url{https://github.com/ocbe-uio/cellmigRation/}
  \item Report bugs at \url{https://github.com/ocbe-uio/cellmigRation/issues}
=======
    \item \url{https://github.com/ocbe-uio/cellmigRation/}
    \item Report bugs at \url{https://github.com/ocbe-uio/cellmigRation/issues}
>>>>>>> BiocNewVersion
}

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no};
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
