% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{cntrd}
\alias{cntrd}
\title{Calculates Centroids}
\usage{
cntrd(im, mx, sz, interactive = NULL)
}
\arguments{
\item{im}{numeric matrix corresponding to the image to process}

\item{mx}{location of local maxima to pixel-levels accuracy}

\item{sz}{diameter of the window over which to average to calculate
the centroid. should be big enough.}

\item{interactive}{numeric; if set to 1 (or any positive number),
an image showing the
computed centroids will be visualized}
}
\value{
a data.frame with 4 columns, containing, x, y, brightness,
and
the square of the radius of gyration for each cell.
}
\description{
Calculates the centroid of bright spots to sub-pixel accuracy.
Inspired by Grier &
Crocker's feature for IDL, but greatly simplified and optimized for
MATLAB, and then
further ported to R.
CREATED: Eric R. Dufresne, Yale University, Feb 4 2005.
}
\examples{
message("the following example may take up to one minute to run")
\dontrun{
x0 <- cellmigRation::TrackCellsDataset@images$images[[1]][100:500, 200:600]
b <- cellmigRation:::bpass(image_array = x0,
                           lnoise = 1,
                           lobject = 15,
                           threshold = 10)
pk <- cellmigRation:::pkfnd(b, 10, 17)
cnt <- cellmigRation:::cntrd(im = b, mx = pk, sz = 17)
cnt[1:10,]
}

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
