% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{DiRatio.Plot}
\alias{DiRatio.Plot}
\title{Directionality Ratio plots}
\usage{
DiRatio.Plot(object, TimeInterval = 10, ExpName = ExpName, export = FALSE)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{export}{if `TRUE` (default), exports plot to JPG file}
}
\value{
Directionality Ratio plots
}
\description{
Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straightline length between the start point and the endpoint of the migration trajectory,
}
\details{
Directionality Ratio
}
\examples{
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
rmTD <-DiRatio(rmTD, ExpName="Test")
DiRatio.Plot(rmTD, ExpName="Test",export=FALSE)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
