% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Directionality.R
\name{DiRatio.Plot}
\alias{DiRatio.Plot}
\title{Directionality Ratio plots}
\usage{
DiRatio.Plot(object, TimeInterval = 10, ExpName = ExpName)
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}
}
\value{
Directionality Ratio plots
}
\description{
Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straightline length between the start point and the endpoint of the migration trajectory,
}
\details{
Method DR

 Directionality Ratio
}
\examples{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
DR=DiRatio(prepro, ExpName="Test")
DiRatio.Plot(prepro, ExpName="Test")

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
