% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/Directionality.R
\name{DiRatio}
\alias{DiRatio}
\title{Directionality Table}
\usage{
DiRatio(object, TimeInterval = 10, ExpName = "ExpName")

DiRatio(object, TimeInterval = 10, ExpName = "ExpName")
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}
}
\value{
An CellMig class object with a data frame stored in the DRtable slot. It contains nine rows: "Cell Number","Directionality Ratio","Mean Cumulative Directionality Ratio","Stable Directionality Ratio", "Number of returns","Min CumDR","Location of Min CumDR, Steps with less CumDR than DR","Directional Persistence".

A data frame with nine rows: "Cell Number","Directionality Ratio","Mean Cumulative Directionality Ratio","Stable Directionality Ratio", "Number of returns","Min CumDR","Location of Min CumDR, Steps with less CumDR than DR","Directional Persistence".
}
\description{
Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straightline length between the start point and the endpoint of the migration trajectory,

Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straightline length between the start point and the endpoint of the migration trajectory,
}
\details{
Method DR

 Directionality Ratio and Directional persistence

Method DR

 Directionality Ratio and Directional persistence
}
\examples{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
rmTD <- rmPreProcessing(rmTD)
rmTD <- DiRatio(rmTD, ExpName="Test")

data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
DR=DiRatio(prepro, ExpName="Test")

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
