% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{DiRatio}
\alias{DiRatio}
\title{Directionality Table}
\usage{
DiRatio(object, TimeInterval = 10, ExpName = "ExpName")
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}
}
\value{
An CellMig class object with a data frame stored in the DRtable slot. It contains nine rows: "Cell Number","Directionality Ratio","Mean Cumulative Directionality Ratio","Stable Directionality Ratio", "Number of returns","Min CumDR","Location of Min CumDR, Steps with less CumDR than DR","Directional Persistence".
}
\description{
Directionality Ratio is the displacement divided by the total length of the total path distance, where displacement is the straightline length between the start point and the endpoint of the migration trajectory,
}
\details{
Directionality Ratio and Directional persistence
}
\examples{
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
rmTD <- DiRatio(rmTD, ExpName="Test")

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
