% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{FinRes}
\alias{FinRes}
\title{Final Results}
\usage{
FinRes(object, ExpName = "ExpName", ParCor = TRUE)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{ParCor}{A logical vector that allows generating a correlation table. Default is TRUE.}
}
\value{
A data frame that contains all the results.
}
\description{
The FinRes function automatically generates a data frame that contains all the results.
}
\details{
Method Final Results
}
\examples{
\dontrun{
data(WSAdataset)
wasDF=WSAdataset[1:2000,]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=95)
wsaTD <-FMI(wsaTD,TimeInterval=10,ExpName="ExpName")
wsaTD <-ForwardMigration(wsaTD,TimeInterval=10,ExpName="ExpName")
wsaTD <-FinRes(wsaTD,ExpName="ExpName",ParCor=FALSE)
}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
