% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{FMI}
\alias{FMI}
\title{Forward Migration Index}
\usage{
FMI(object, TimeInterval = 10, ExpName = "ExpName")
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}
}
\value{
An CellMig class Object with a data frame. The data frame is stored in the FMItable slot.
}
\description{
The FMI function automatically generates data for the forward migration index
}
\details{
Method Forward Migration Index
}
\examples{
data(WSAdataset)
wasDF=WSAdataset[1:2000,]
wsaTD <- CellMig(wasDF)
wsaTD <- wsaPreProcessing(wsaTD,FrameN=95)
wsaTD <-FMI(wsaTD,TimeInterval=10,ExpName="ExpName")

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
