% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{getCellImages}
\alias{getCellImages}
\alias{getCellImages,trackedCells-method}
\title{Method getCellImages}
\usage{
getCellImages(x)

\S4method{getCellImages}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\description{
Retrieve Images from a \code{trackedCells} object.
}
