% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getCellMigSlot}
\alias{getCellMigSlot}
\alias{getCellMigSlot,CellMig,character-method}
\title{Method getCellMigSlot}
\usage{
getCellMigSlot(x, slot)

\S4method{getCellMigSlot}{CellMig,character}(x, slot)
}
\arguments{
\item{x}{a \code{CellMig}-class object}

\item{slot}{string pointing to the slot to be retrieved}
}
\value{
a slot from a CellMig object
}
\description{
Get Data from a slot in a \code{CellMig} object.
}
\examples{
data("TrajectoryDataset")
x <- CellMig(TrajectoryDataset)
getCellMigSlot(x, "trajdata")

}
