% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{getCellsStats}
\alias{getCellsStats}
\title{Get Cell migration stats}
\usage{
getCellsStats(tc_obj)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}
}
\value{
data.frame including cell migration stats
}
\description{
Extract cell migration statistics from a trackedCells object
}
\examples{
message("this example may take up to several mins to complete")
\dontrun{
x <- cellmigRation::TrackCellsDataset
x <- OptimizeParams(x)
x <- CellTracker(x)
x <- ComputeTracksStats(x, 10, 10)
getCellsStats(x)[1:10,]
}



}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
