% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getImageCentroids}
\alias{getImageCentroids}
\alias{getImageCentroids,trackedCells-method}
\title{Method getImageCentroids}
\usage{
getImageCentroids(x)

\S4method{getImageCentroids}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a list including all centroids
}
\description{
Retrieve Image Centroids from a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
getImageCentroids(TrackCellsDataset)

}
