% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/all_functions.R
\name{getOptimizedParams}
\alias{getOptimizedParams}
\title{Get Auto Optimized Parameters}
\usage{
getOptimizedParams(tc_obj)
}
\arguments{
\item{tc_obj}{a \code{trackedCells} object}
}
\value{
a list including optimized parameter values
(lnoise, diameter, and threshold)
}
\description{
Extract Parameters that were automatically optimized
}
\examples{
message("this example may take up to several mins to complete")
\dontrun{
x <- cellmigRation::TrackCellsDataset
x <- OptimizeParams(x)
getOptimizedParams(x)
}


}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
=======
% Please edit documentation in R/all_methods_v3.R
\name{getOptimizedParams}
\alias{getOptimizedParams}
\alias{getOptimizedParams,trackedCells-method}
\title{Method getOptimizedParams}
\usage{
getOptimizedParams(x)

\S4method{getOptimizedParams}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a list including Optimized Parameters
}
\description{
Retrieve Optimized Params from a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
getOptimizedParams(TrackCellsDataset)

>>>>>>> BiocNewVersion
}
