% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{getOptimizedParams}
\alias{getOptimizedParams}
\alias{getOptimizedParams,trackedCells-method}
\title{Method getOptimizedParams}
\usage{
getOptimizedParams(x)

\S4method{getOptimizedParams}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\description{
Retrieve Optimized Params from a \code{trackedCells} object.
}
