% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{getOptimizedParams}
\alias{getOptimizedParams}
\alias{getOptimizedParams,trackedCells-method}
\title{Method getOptimizedParams}
\usage{
getOptimizedParams(x)

\S4method{getOptimizedParams}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\value{
a list including Optimized Parameters
}
\description{
Retrieve Optimized Params from a \code{trackedCells} object.
}
<<<<<<< HEAD
\examples{
data("TrackCellsDataset")
getOptimizedParams(TrackCellsDataset)

}
