% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions_v8.R
\name{getPerAndSpeed}
\alias{getPerAndSpeed}
\title{Getting the table of Persistence and Speed.}
\usage{
getPerAndSpeed(object)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of
data frames resulted from the PreProcessing.}
}
\value{
A data frame of Persistence and Speed.
}
\description{
The PerAndSpeed() generates data and plots for
persistence and speed.
}
\examples{
rmTD <- cellmigRation::preProcCellMig
rmTD <- PerAndSpeed(rmTD,TimeInterval=10, export=FALSE)
head(getPerAndSpeed(rmTD))

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
