% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{getProcessedImages}
\alias{getProcessedImages}
\alias{getProcessedImages,trackedCells-method}
\title{Method getProcessedImages}
\usage{
getProcessedImages(x)

\S4method{getProcessedImages}{trackedCells}(x)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}
}
\description{
Retrieve Processed Images from a \code{trackedCells} object.
}
