% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerAndSpeed.R
\name{PerAndSpeed}
\alias{PerAndSpeed}
\title{Persistence and Speed}
\usage{
PerAndSpeed(object, TimeInterval = 10, ExpName = "ExpName",
  PtSplot = TRUE, AllPtSplot = TRUE, ApSplot = TRUE,
  AllApSplot = TRUE)
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{PtSplot}{A logical vector that allows generating individual plots of persistence time vs speed per cell. Default is TRUE.}

\item{AllPtSplot}{A logical vector that allows generating a plot of persistence time vs speed for all cells. Default is TRUE.}

\item{ApSplot}{A logical vector that allows generating individual plots of angular persistence vs speed per cell. Default is TRUE.}

\item{AllApSplot}{A logical vector that allows generating a plot of angular persistence vs speed of all cells. Default is TRUE.}
}
\value{
A data frame and plots
}
\description{
The PerAndSpeed() generates data and plots for persistence and speed.
}
\details{
Method Persistence and Speed
}
\examples{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
PerSp<-PerAndSpeed(prepro,TimeInterval=10,ExpName="ExpName")


}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
