% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/Plotting.R
\name{plot3DAllTracks}
\alias{plot3DAllTracks}
\title{A 3D rose-plot of all cells}
\usage{
plot3DAllTracks(object, VS = 3, size = 2)

plot3DAllTracks(object, VS = 3, size = 2)
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{VS}{A numeric value of the vertical separator between cells.}

\item{size}{A numeric value of the point's size.}
}
\value{
A 3D rose-plot showing the tracks of all cells.

A 3D rose-plot showing the tracks of all cells.
}
\description{
Plotting the trajectory data of all cells in 3D.

Plotting the trajectory data of all cells in 3D.
}
\details{
The 3D visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).

The 3D visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
rmTD <- rmPreProcessing(rmTD)
plot3DAllTracks(rmTD, VS=3, size=2)


data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
plot3DAllTracks(prepro, VS=3, size=2)

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
