% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/Plotting.R
\name{plot3DTracks}
\alias{plot3DTracks}
\title{A 3D rose-plot}
\usage{
plot3DTracks(object, VS = 3, size = 2, cells)

plot3DTracks(object, VS = 3, size = 2, cells)
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{VS}{A numeric value of the vertical separator between cells.}

\item{size}{A numeric value of the point's size.}

\item{cells}{A numeric vector containing the cell's numbers to be plotted.}
}
\value{
A 3D rose-plot showing the tracks of particular cells.

A 3D rose-plot showing the tracks of particular cells.
}
\description{
Plotting the trajectory data of particular cells in 3D.

Plotting the trajectory data of particular cells in 3D.
}
\details{
The 3D visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).

The 3D visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
\dontrun{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
rmTD <- rmPreProcessing(rmTD)
plot3DTracks(rmTD, VS=3, size=2,cells=c(1,50,150,250,350))
}

\dontrun{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
plot3DTracks(prepro, VS=3, size=2,cells=c(1,50,150,250,350))
}

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
