% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R
\name{plotAllTracks}
\alias{plotAllTracks}
\title{A 2D rose-plot}
\usage{
plotAllTracks(object, ExpName = "ExpName", Type = "l")
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{Type}{has to be one of the following: [p, l, b, o]
“p”: Points
“l”: Lines
“b”: Both
“o”: Both “overplotted”}
}
\value{
A 2D rose-plot showing the tracks of all cells.
}
\description{
Plotting the trajectory data of all cells.
}
\details{
Method Plotting

 The visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(Trajectory_dataset)
rmDF=Trajectory_dataset[1:2000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
plotAllTracks(rmTD, ExpName="Test",Type="b")


}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
