% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMig.R, R/Plotting.R
\name{plotAllTracks}
\alias{plotAllTracks}
\title{A 2D rose-plot}
\usage{
plotAllTracks(object, ExpName = "ExpName", Type = "l")

plotAllTracks(object, ExpName = "ExpName", Type = "l")
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{Type}{has to be one of the following: [p, l, b, o]
“p”: Points
“l”: Lines
“b”: Both
“o”: Both “overplotted”}
}
\value{
A 2D rose-plot showing the tracks of all cells.

A 2D rose-plot showing the tracks of all cells.
}
\description{
Plotting the trajectory data of all cells.

Plotting the trajectory data of all cells.
}
\details{
Method Plotting

 The visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).

Method Plotting

 The visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(Trajectory_dataset)
rmTD <- CellMig(Trajectory_dataset)
\dontrun{
rmTD <- rmPreProcessing(rmTD)
plotAllTracks(rmTD, ExpName="Test",Type="b")
}


\dontrun{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
plotAllTracks(prepro, ExpName="Test",Type="b")
}

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}

Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
