% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{PlotTracksSeparately}
\alias{PlotTracksSeparately}
\title{A graphical display of the track of each cell.}
\usage{
PlotTracksSeparately(
  object,
  ExpName = "ExpName",
  Type = "l",
  FixedField = TRUE,
  export = TRUE
)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{Type}{has to be one of the following: [p, l, b, o]
"p": Points
"l": Lines
"b": Both
"o": Both "overplotted"}

\item{FixedField}{logical(1) Allows generating individual plots with fixed field. Default is TRUE.}

\item{export}{if `TRUE` (default), exports plot to JPG file}
}
\value{
2D rose-plots of the cells' track Separately.
}
\description{
Plotting the trajectory data of each cell.
}
\details{
The visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
PlotTracksSeparately(rmTD, ExpName="Test",
                     Type="b", FixedField=FALSE, export = FALSE)

}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
