% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotTracksSeparately}
\alias{PlotTracksSeparately}
\title{A graphical display of the track of each cell.}
\usage{
PlotTracksSeparately(object, ExpName = "ExpName", Type = "l",
  FixedField = TRUE)
}
\arguments{
\item{object}{A list of data frames resulted from runnning the function "PreProcessing()".}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{Type}{has to be one of the following: [p, l, b, o]
“p”: Points
“l”: Lines
“b”: Both
“o”: Both “overplotted”}

\item{FixedField}{logical(1) Allows generating individual plots with fixed field. Default is TRUE.}
}
\value{
2D rose-plots of the cells' track Separately.
}
\description{
Plotting the trajectory data of each cell.
}
\details{
The visualization shows centered trajectories where the starting point of each track is located at the origin of the coordinate system (X=0,Y=0).
}
\examples{
data(Trajectory_dataset)
df<-Trajectory_dataset
prepro<-PreProcessing(df,PixelSize=1.24, TimeInterval=10)
PlotTracksSeparately(prepro, ExpName="Test",Type="b",FixedField=FALSE)

}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
