% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{preProcCellMig}
\alias{preProcCellMig}
\title{Trajectories of 11 cells}
\format{
a list including 21 elements
}
\usage{
data(preProcCellMig)
}
\description{
Intermediates and Results from Cell Tracking Analyses,
used as a representative example of a S4 CellMig object
}
\details{
BT549 cell trajectories were computed by using cellmigRation.  
Imaging experiments were performed as described by Ghannoum S 
et al (paper in preparation). Briefly, triple negative breast 
cancer BT549 cells were cultured in RPMI supplemented with 10% FCS 
and 1% penicillin/streptomycin. Cells were transduced with 
NucLight green lentivirus (Essen BioScience), and then sorted by  
fluorescence-activated cell sorting (FACS). GFP-positive cells 
were seeded at a 1:3 ratio with untransduced BT549 cells in 
96-well image-lock plates (EssenBio) at a density of 1000 
total cells per well. Once cells reached the desired density, 
they were scanned at ten-minute intervals over 24h using an 
Incucyte S3 Live-Cell microscope (EssenBio) at 10x magnification 
and a Basler Ace 1920-155um camera with CMOS sensor. 
TIFF images 
were imported and processed using the cellmigRation library.
}
\examples{
data(preProcCellMig)

}
\keyword{internal}
