% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{setAnalyticParams}
\alias{setAnalyticParams}
\alias{setAnalyticParams,trackedCells,list-method}
\title{Method setAnalyticParams}
\usage{
setAnalyticParams(x, params)

\S4method{setAnalyticParams}{trackedCells,list}(x, params)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{params}{a list including all params}
}
\description{
Set Analytic Params of a \code{trackedCells} object.
}
