% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{setOptimizedParams}
\alias{setOptimizedParams}
\alias{setOptimizedParams,trackedCells-method}
\alias{setOptimizedParams,trackedCells,ANY,ANY-method}
\title{Method setOptimizedParams}
\usage{
setOptimizedParams(x, auto_params, results)

\S4method{setOptimizedParams}{trackedCells}(x, auto_params, results)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{auto_params}{automatically selected parameters}

\item{results}{optimization analysis results}
}
\description{
Set Optimized Params of a \code{trackedCells} object.
}
