% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\name{setProcessedImages}
\alias{setProcessedImages}
\alias{setProcessedImages,trackedCells,list-method}
\title{Method setProcessedImages}
\usage{
setProcessedImages(x, procImages)

\S4method{setProcessedImages}{trackedCells,list}(x, procImages)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{procImages}{a list including all metadata}
}
\description{
Set Processed Images of a \code{trackedCells} object.
}
