% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setProcessingStatus}
\alias{setProcessingStatus}
\alias{setProcessingStatus,trackedCells,character,numeric-method}
\title{Method setProcessingStatus}
\usage{
setProcessingStatus(x, slot, value)

\S4method{setProcessingStatus}{trackedCells,character,numeric}(x, slot, value)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{slot}{string pointing to the slot to be updated}

\item{value}{numeric value to be written}
}
\value{
a trackedCells object
}
\description{
Set Operation Status of a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
setProcessingStatus(TrackCellsDataset, slot="optimized_params", value=0)


}
