% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setTrackedCellsMeta}
\alias{setTrackedCellsMeta}
\alias{setTrackedCellsMeta,trackedCells,list-method}
\title{Method setTrackedCellsMeta}
\usage{
setTrackedCellsMeta(x, meta)

\S4method{setTrackedCellsMeta}{trackedCells,list}(x, meta)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{meta}{a list including all metadata}
}
\value{
a trackedCells object
}
\description{
Set Metadata of a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
meta <- getCellTrackMeta(TrackCellsDataset)
meta[["condition"]] <- "DEMO N.2"
setTrackedCellsMeta(TrackCellsDataset, meta = meta)

}
