% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods_v3.R
\name{setTrackedPositions}
\alias{setTrackedPositions}
\alias{setTrackedPositions,trackedCells,data.frame-method}
\title{Method setTrackedPositions}
\usage{
setTrackedPositions(x, positions)

\S4method{setTrackedPositions}{trackedCells,data.frame}(x, positions)
}
\arguments{
\item{x}{a \code{trackedCells}-class object}

\item{positions}{a data.frame including all positions}
}
\value{
a trackedCells object
}
\description{
Set positions of a \code{trackedCells} object.
}
\examples{
data("TrackCellsDataset")
setTrackedPositions(TrackCellsDataset, data.frame())

}
