% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{sinkAway}
\alias{sinkAway}
\title{Sinking Output as part of Parameter Optimization 
 
Tool for Sinking Output as part of the Optimization Parameter process}
\usage{
sinkAway(ty = "M", fl = "/dev/null")
}
\arguments{
\item{ty}{character, "M" for messages, anything else for output}

\item{fl}{filename to sink the output to}
}
\value{
value returned by the sink() function
}
\description{
Sinking Output as part of Parameter Optimization 
 
Tool for Sinking Output as part of the Optimization Parameter process
}
\details{
This is an internal function supporting the Optimization 
Parameter process
}
\examples{
print(1)
cellmigRation:::sinkAway("O")
print(2)
cellmigRation:::sinkAway("O", NULL)
print(3)

}
\keyword{internal}
