% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{ThreeConditions}
\alias{ThreeConditions}
\title{Intermediates and Results from Cell Tracking Analyses}
\format{
a list including 3 elements
}
\usage{
data(ThreeConditions)
}
\description{
Intermediates and Results from Cell Tracking Analyses,
used to build the package vignette.
}
\examples{
data(ThreeConditions)

}
\keyword{internal}
