% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_extra.R
\docType{data}
\name{TrackCellsDataset}
\alias{TrackCellsDataset}
\title{Sample Stack of Fluorescent Cells}
\format{
a \code{trackedCells} object including 10 stacks
}
\usage{
data(TrackCellsDataset)
}
\description{
Sample Stack of Fluorescent Cells to be used for
computing
cell tracks and stats
}
\examples{
data(TrackCellsDataset)

}
\keyword{internal}
