% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{VeAutoCor}
\alias{VeAutoCor}
\title{Velocity AutoCorrelation}
\usage{
VeAutoCor(
  object,
  TimeInterval = 10,
  ExpName = "ExpName",
  sLAG = 0.25,
  sPLOT = TRUE,
  aPLOT = TRUE,
  export = FALSE
)
}
\arguments{
\item{object}{\code{CellMig} class object, which is a list of data frames resulted from the PreProcessing.}

\item{TimeInterval}{A numeric value of the time elapsed between successive frames in the time-lapse stack.}

\item{ExpName}{A character string. The ExpName will be appended to all exported tracks and statistics data}

\item{sLAG}{A numeric value to be used to get the number of lags for the slope fitting. Default is 0.25, which represents 25 percent of the steps.}

\item{sPLOT}{A logical vector that allows generating individual plots showing the velocity across several sequantial time intervals. Default is TRUE.}

\item{aPLOT}{A logical vector that allows generating a plot showing the velocity across several sequantial time intervals of all cells. Default is TRUE.}

\item{export}{if `TRUE` (default), exports function output to CSV file}
}
\value{
Plots and a data frame, which contains six rows: "Cell Number", .
}
\description{
The VeAutoCor function automatically compute the changes in both speed and direction across several sequantial time intervals.
}
\examples{
\dontrun{
data(TrajectoryDataset)
rmDF=TrajectoryDataset[1:1000,]
rmTD <- CellMig(rmDF)
rmTD <- rmPreProcessing(rmTD,FrameN=100)
rmTD <- VeAutoCor(
   rmTD, TimeInterval=10, ExpName="ExpName", sLAG=0.25, sPLOT=TRUE,
   aPLOT=TRUE, export=FALSE
)
}
}
\references{
\url{https://www.data-pulse.com/dev_site/cellmigration/}
}
\author{
Salim Ghannoum \email{salim.ghannoum@medisin.uio.no}
}
