% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSAdataset.R
\docType{data}
\name{WSAdataset}
\alias{WSAdataset}
\title{Trajectories of 147 cells}
\format{A data frame with 11970 rows and 4 columns}
\usage{
data(WSAdataset)
}
\description{
A dataset containing the coordinates and the ID of 147 cells from wound scratch migration experiment
}
\examples{
data(WSAdataset)

}
\keyword{datasets}
